Test improvements. More minor cleanups.
git-svn-id: https://crosswire.org/svn/sword/trunk@3389 bcd7d363-81e1-0310-97ec-a550e20fc99c

+++ bindings/objc/src/SwordVerseKey.mm
@@ -67,7 +67,7 @@
         if(scheme) {
             [self setVersification:scheme];
         }
-        
+
         if(aRef) {
             [self setKeyText:aRef];
         }
@@ -82,7 +82,7 @@
     return [SwordVerseKey verseKeyWithSWVerseKey:(sword::VerseKey *)sk];
 }
 
-- (int)index {
+- (long)index {
     return ((sword::VerseKey *)sk)->getIndex();
 }
